Refactor search service into common
