+++ src/api/services/map/definition.ts
@@ -2,7 +2,7 @@ import * as zod from "zod";
 import { createTagFactory } from "../../../lib/createTagFactory";
 import { createRpcDefinition } from "../../util/rpc";
 import { UserAccessLevel } from "../auth/types";
-import { createSearchTypes } from "../search/types";
+import { createSearchTypes } from "../../common/search";
 import {
   mapIdType,
   mapInfoFilter,
