+++ src/api/services/monster/definition.ts
@@ -1,6 +1,6 @@
 import { createTagFactory } from "../../../lib/createTagFactory";
 import { createRpcDefinition } from "../../util/rpc";
-import { createSearchTypes } from "../search/types";
+import { createSearchTypes } from "../../common/search";
 import {
   monsterFilter,
   monsterSpawnFilter,
